ensure updated time is in correct time zone
With ruby 1.9.2 and greater makes sure the updated time is in London
local time, whatever the environment local time is.

With older versions of ruby time will be correct, but in environment
local time, so if you're in London everything works fine, but if you
have a server in San Francisco "16:30 +01:00" will end up as
"10:30 -07:00".